Undo help.yaml changes in #8325, #8503 and #8535
