+++ src/azure-cli-core/azure/cli/core/tests/test_help.py
@@ -7,14 +7,14 @@ from __future__ import print_function
 
 import logging
 import shutil
-import inspect
-from inspect import getmembers as inspect_getmembers
+from knack.help import GroupHelpFile, HelpAuthoringException
+
 
 import unittest
 import mock
 import tempfile
 
-from knack.help import GroupHelpFile, HelpAuthoringException
+from azure.cli.core import AzCommandsLoader
 from azure.cli.core._help import CliCommandHelpFile
 
 from azure.cli.core.mock import DummyCli
@@ -22,37 +22,6 @@ from azure.cli.core.commands import _load_command_loader
 from azure.cli.core.file_util import get_all_help
 
 logger = logging.getLogger(__name__)
-
-# Command loader module
-MOCKED_COMMAND_LOADER_MOD = "test_help_loaders"
-
-
-# mock command loader method so that only the mock command loader can be loaded.
-def mock_load_command_loader(loader, args, name, prefix):
-    return _load_command_loader(loader, args, name, "azure.cli.core.tests.")
-
-
-# mock inspect.getfile to discover directory containing help files.
-def get_mocked_inspect_getfile(expected_arg, return_value):
-
-    def inspect_getfile(obj):
-        if obj == expected_arg:
-            return return_value
-        else:
-            return inspect.getfile(obj)
-    return inspect_getfile
-
-
-# mock getmembers so test help loader can be injected into AzCliHelp class' versioned loader list
-def mock_inspect_getmembers(object, predicate=None):
-    import azure.cli.core.tests.test_help_loaders as possible_loaders
-
-    if "azure.cli.core._help_loaders" in repr(object) and "is_loader_cls" in repr(predicate):
-        result = inspect_getmembers(object, predicate)
-        result.extend(inspect_getmembers(possible_loaders, predicate))
-        return list(set(result))
-    else:
-        return inspect_getmembers(object, predicate)
 
 
 def _store_parsers(parser, d):
@@ -71,7 +40,6 @@ def _is_group(parser):
 def _get_parser_name(parser):
     # pylint:disable=protected-access
     return (parser._prog_prefix if hasattr(parser, '_prog_prefix') else parser.prog)[len('az '):]
-
 
 def create_invoker_and_load_cmds_and_args(cli_ctx):
     from knack import events
@@ -108,8 +76,54 @@ def create_invoker_and_load_cmds_and_args(cli_ctx):
     cli_ctx.raise_event(events.EVENT_INVOKER_POST_CMD_TBL_CREATE, commands_loader=invoker.commands_loader)
     invoker.parser.load_command_table(invoker.commands_loader)
 
+# Command loader module
+MOCKED_COMMAND_LOADER_MOD = "test_help"
 
-# TODO update this CLASS to properly load all help...                                                                       .
+# mock command loader method so that only the mock command loader can be loaded.
+def mock_load_command_loader(loader, args, name, prefix):
+    return _load_command_loader(loader, args, name, "azure.cli.core.tests.")
+
+# region TestCommandLoader
+class TestCommandLoader(AzCommandsLoader):
+
+    def __init__(self, cli_ctx=None):
+        from azure.cli.core.commands import CliCommandType
+        compute_custom = CliCommandType(
+            operations_tmpl='{}#{{}}'.format(__name__),
+        )
+        super(TestCommandLoader, self).__init__(cli_ctx=cli_ctx,
+                                                custom_command_type=compute_custom)
+        self.cmd_to_loader_map = {}
+
+    def load_command_table(self, args):
+        with self.command_group('test') as g:
+            g.custom_command('alpha', 'dummy_handler')
+
+        return self.command_table
+
+    def load_arguments(self, command):
+        with self.argument_context('test') as c:
+            c.argument('arg1', options_list=['--arg1', '-a'])
+            c.argument('arg2', options_list=['--arg2', '-b'], help="Help From code.")
+        with self.argument_context('test alpha') as c:
+            c.positional('arg4', metavar="ARG4")
+        self._update_command_definitions()  # pylint: disable=protected-access
+
+
+def dummy_handler(arg1, arg2=None, arg3=None, arg4=None):
+    """
+    Short summary here. Long summary here. Still long summary.
+    :param arg1: arg1's docstring help text
+    :param arg2: arg2's docstring help text
+    :param arg3: arg3's docstring help text
+    :param arg4: arg4's docstring help text
+    """
+    pass
+
+COMMAND_LOADER_CLS = TestCommandLoader
+# endregion
+
+
 class HelpTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
@@ -196,130 +210,6 @@ class TestHelpLoads(unittest.TestCase):
                   unsupported-profiles: 2017-03-09-profile
         """
 
-    def set_help_yaml(self):
-        yaml_help = """
-        version: 1
-        content:
-        - group:
-            name: test
-            summary: Group yaml summary
-            description: Group yaml description. A.K.A long description
-            links:
-                - title: Azure Test Docs
-                  url: "https://docs.microsoft.com/en-us/azure/test"
-                - url: "https://aka.ms/just-a-url"
-        - command:
-            name: test alpha
-            summary: Command yaml summary
-            description: Command yaml description. A.K.A long description
-            links:
-                - title: Azure Test Alpha Docs
-                  url: "https://docs.microsoft.com/en-us/azure/test/alpha"
-                - url: "https://aka.ms/just-a-long-url"
-            arguments:
-                - name: --arg2 # we do not specify the short option in the name.
-                  summary: Arg 2's summary
-                  description: A true description of this parameter.
-                  value-sources:
-                    - string: "Number range: -5.0 to 5.0"
-                    - link:
-                        url: https://www.foo.com
-                        title: foo
-                    - link:
-                        command: az test show
-                        title: Show test details
-                - name: ARG4  # Note: positional's are discouraged in the CLI.
-                  summary: Arg4's summary, yaml. Positional arg, not required
-            examples:
-                - summary: A simple example
-                  description: More detail on the simple example.
-                  command: az test alpha --arg1 apple --arg2 ball --arg3 cat
-                  supported-profiles: 2018-03-01-hybrid, latest
-                - summary: Another example unsupported on latest
-                  description: More detail on the unsupported example.
-                  command: az test alpha --arg1 apple --arg2 ball
-                  unsupported-profiles: 2017-03-09-profile
-        """
-        return self._create_new_temp_file(yaml_help, suffix="help.yaml")
-
-    def set_help_json(self):
-        json_help = """
-            {
-                "version": 2,
-                "content": [
-                    {
-                        "group": {
-                            "name": "test",
-                            "short": "Group json summary",
-                            "long": "Group json description. A.K.A long description",
-                            "hyper-links": [
-                                {
-                                    "title": "Azure Json Test Docs",
-                                    "url": "https://docs.microsoft.com/en-us/azure/test"
-                                },
-                                {
-                                    "url": "https://aka.ms/just-a-url"
-                                }
-                            ]
-                        }
-                    },
-                    {
-                        "command": {
-                            "name": "test alpha",
-                            "short": "Command json summary",
-                            "long": "Command json description. A.K.A long description",
-                            "hyper-links": [
-                                {
-                                    "title": "Azure Json Test Alpha Docs",
-                                    "url": "https://docs.microsoft.com/en-us/azure/test/alpha"
-                                },
-                                {
-                                    "url": "https://aka.ms/just-a-long-url"
-                                }
-                            ],
-                            "arguments": [
-                                {
-                                    "name": "--arg3",
-                                    "short": "Arg 3's json summary",
-                                    "long": "A truly true description of this parameter.",
-                                    "sources": [
-                                        {
-                                            "string": "Number range: 0 to 10"
-                                        },
-                                        {
-                                            "link": {
-                                                "url": "https://www.foo-json.com",
-                                                "title": "foo-json"
-                                            }
-                                        },
-                                        {
-                                            "link": {
-                                                "command": "az test show",
-                                                "title": "Show test details. Json file"
-                                            }
-                                        }
-                                    ]
-                                },
-                                {
-                                    "name": "ARG4",
-                                    "summary": "Arg4's summary, json. Positional arg, not required"
-                                }
-                            ],
-                            "examples": [
-                                {
-                                    "summary": "A simple example from json",
-                                    "description": "More detail on the simple example.",
-                                    "command": "az test alpha --arg1 alpha --arg2 beta --arg3 chi",
-                                    "supported-profiles": "2018-03-01-hybrid, latest"
-                                }
-                            ]
-                        }
-                    }
-                ]
-            }
-        """
-        return self._create_new_temp_file(json_help, suffix="help.json")
-
     # Mock logic in core.MainCommandsLoader.load_command_table for retrieving installed modules.
     @mock.patch('pkgutil.iter_modules', side_effect=lambda x: [(None, MOCKED_COMMAND_LOADER_MOD, None)])
     @mock.patch('azure.cli.core.commands._load_command_loader', side_effect=mock_load_command_loader)
@@ -362,152 +252,16 @@ class TestHelpLoads(unittest.TestCase):
         self.assertEqual(obj_param_dict["--arg1 -a"].short_summary, "A short summary.")
 
         self.assertEqual(obj_param_dict["--arg1 -a"].short_summary, "A short summary.")
-        self.assertEqual(obj_param_dict["--arg1 -a"].value_sources[0]['link']['command'], "az foo bar")
-        self.assertEqual(obj_param_dict["--arg1 -a"].value_sources[1]['link']['command'], "az bar baz")
+        self.assertEqual(obj_param_dict["--arg1 -a"].value_sources[0], "az foo bar")
+        self.assertEqual(obj_param_dict["--arg1 -a"].value_sources[1], "az bar baz")
 
-        self.assertEqual(command_help_obj.examples[0].short_summary, "Alpha Example")
-        self.assertEqual(command_help_obj.examples[0].command, "az test alpha --arg1 a --arg2 b --arg3 c")
+        self.assertEqual(command_help_obj.examples[0].name, "Alpha Example")
+        self.assertEqual(command_help_obj.examples[0].text, "az test alpha --arg1 a --arg2 b --arg3 c")
         self.assertEqual(command_help_obj.examples[0].supported_profiles, "2018-03-01-hybrid, latest")
         self.assertEqual(command_help_obj.examples[0].unsupported_profiles, None)
 
         self.assertEqual(command_help_obj.examples[1].supported_profiles, None)
         self.assertEqual(command_help_obj.examples[1].unsupported_profiles, "2017-03-09-profile")
-
-    @mock.patch('pkgutil.iter_modules', side_effect=lambda x: [(None, MOCKED_COMMAND_LOADER_MOD, None)])
-    @mock.patch('azure.cli.core.commands._load_command_loader', side_effect=mock_load_command_loader)
-    def test_load_from_help_yaml(self, mocked_load, mocked_pkg_util):
-        # setup help.py and help.yaml help.
-        self.set_help_py()
-        yaml_path = self.set_help_yaml()
-        create_invoker_and_load_cmds_and_args(self.test_cli)
-
-        # mock logic in core._help_loaders for retrieving yaml file from loader path.
-        expected_arg = self.test_cli.invocation.commands_loader.cmd_to_loader_map['test alpha'][0].__class__
-        with mock.patch('inspect.getfile', side_effect=get_mocked_inspect_getfile(expected_arg, yaml_path)):
-            group_help_obj = next((help for help in get_all_help(self.test_cli) if help.command == "test"), None)
-            command_help_obj = next((help for help in get_all_help(self.test_cli) if help.command == "test alpha"), None)  # pylint: disable=line-too-long
-
-            # Test that group and command help are successfully displayed.
-            with self.assertRaises(SystemExit):
-                self.test_cli.invoke(["test", "-h"])
-            with self.assertRaises(SystemExit):
-                self.test_cli.invoke(["test", "alpha", "-h"])
-
-        # Test group help
-        self.assertIsNotNone(group_help_obj)
-        self.assertEqual(group_help_obj.short_summary, "Group yaml summary.")
-        self.assertEqual(group_help_obj.long_summary, "Group yaml description. A.K.A long description.")
-        self.assertEqual(group_help_obj.links[0], {"title": "Azure Test Docs", "url": "https://docs.microsoft.com/en-us/azure/test"})
-        self.assertEqual(group_help_obj.links[1], {"url": "https://aka.ms/just-a-url"})
-
-        # Test command help
-        self.assertIsNotNone(command_help_obj)
-        self.assertEqual(command_help_obj.short_summary, "Command yaml summary.")
-        self.assertEqual(command_help_obj.long_summary, "Command yaml description. A.K.A long description.")
-        self.assertEqual(command_help_obj.links[0], {"title": "Azure Test Alpha Docs",
-                                                     "url": "https://docs.microsoft.com/en-us/azure/test/alpha"})
-        self.assertEqual(command_help_obj.links[1], {"url": "https://aka.ms/just-a-long-url"})
-
-        # test that parameters and help are loaded from command function docstring, argument registry help and help.yaml
-        obj_param_dict = {param.name: param for param in command_help_obj.parameters}
-        param_name_set = {"--arg1 -a", "--arg2 -b", "--arg3", "ARG4"}
-        self.assertTrue(set(obj_param_dict.keys()).issuperset(param_name_set))
-
-        self.assertEqual(obj_param_dict["--arg1 -a"].short_summary, "A short summary.")
-        self.assertEqual(obj_param_dict["--arg3"].short_summary, "Arg3's docstring help text.")
-        self.assertEqual(obj_param_dict["ARG4"].short_summary, "Arg4's summary, yaml. Positional arg, not required.")
-
-        self.assertEqual(obj_param_dict["--arg2 -b"].short_summary, "Arg 2's summary.")
-        self.assertEqual(obj_param_dict["--arg2 -b"].long_summary, "A true description of this parameter.")
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[0], {"string": "Number range: -5.0 to 5.0"})
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[1]['link'], {"url": "https://www.foo.com",
-                                                                                "title": "foo"})
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[2]['link'], {"command": "az test show",
-                                                                                "title": "Show test details"})
-
-        self.assertEqual(command_help_obj.examples[0].short_summary, "A simple example")
-        self.assertEqual(command_help_obj.examples[0].long_summary, "More detail on the simple example.")
-        self.assertEqual(command_help_obj.examples[0].command, "az test alpha --arg1 apple --arg2 ball --arg3 cat")
-        self.assertEqual(command_help_obj.examples[0].supported_profiles, "2018-03-01-hybrid, latest")
-        self.assertEqual(command_help_obj.examples[0].unsupported_profiles, None)
-
-        self.assertEqual(command_help_obj.examples[1].supported_profiles, None)
-        self.assertEqual(command_help_obj.examples[1].unsupported_profiles, "2017-03-09-profile")
-
-    @mock.patch('inspect.getmembers', side_effect=mock_inspect_getmembers)
-    @mock.patch('pkgutil.iter_modules', side_effect=lambda x: [(None, MOCKED_COMMAND_LOADER_MOD, None)])
-    @mock.patch('azure.cli.core.commands._load_command_loader', side_effect=mock_load_command_loader)
-    def test_load_from_help_json(self, mocked_load, mocked_pkg_util, mocked_getmembers):
-        # setup help.py, help.yaml and help.json
-        self.set_help_py()
-        path = self.set_help_yaml()  # either (yaml or json) path should work. As both files are in the same temp dir.
-        self.set_help_json()
-        create_invoker_and_load_cmds_and_args(self.test_cli)
-
-        # mock logic in core._help_loaders for retrieving yaml file from loader path.
-        expected_arg = self.test_cli.invocation.commands_loader.cmd_to_loader_map['test alpha'][0].__class__
-        with mock.patch('inspect.getfile', side_effect=get_mocked_inspect_getfile(expected_arg, path)):
-            group_help_obj = next((help for help in get_all_help(self.test_cli) if help.command == "test"), None)
-            command_help_obj = next((help for help in get_all_help(self.test_cli) if help.command == "test alpha"),
-                                    None)
-
-            # Test that group and command help are successfully displayed.
-            with self.assertRaises(SystemExit):
-                self.test_cli.invoke(["test", "-h"])
-            with self.assertRaises(SystemExit):
-                self.test_cli.invoke(["test", "alpha", "-h"])
-
-        # Test group help
-        self.assertIsNotNone(group_help_obj)
-        self.assertEqual(group_help_obj.short_summary, "Group json summary.")
-        self.assertEqual(group_help_obj.long_summary, "Group json description. A.K.A long description.")
-        self.assertEqual(group_help_obj.links[0], {"title": "Azure Json Test Docs",
-                                                   "url": "https://docs.microsoft.com/en-us/azure/test"})
-        self.assertEqual(group_help_obj.links[1], {"url": "https://aka.ms/just-a-url"})
-
-        # Test command help
-        self.assertIsNotNone(command_help_obj)
-        self.assertEqual(command_help_obj.short_summary, "Command json summary.")
-        self.assertEqual(command_help_obj.long_summary, "Command json description. A.K.A long description.")
-        self.assertEqual(command_help_obj.links[0], {"title": "Azure Json Test Alpha Docs",
-                                                     "url": "https://docs.microsoft.com/en-us/azure/test/alpha"})
-        self.assertEqual(command_help_obj.links[1], {"url": "https://aka.ms/just-a-long-url"})
-
-        # test that parameters and help are loaded from command function docstring, argument registry help and help.yaml
-        obj_param_dict = {param.name: param for param in command_help_obj.parameters}
-        param_name_set = {"--arg1 -a", "--arg2 -b", "--arg3", "ARG4"}
-        self.assertTrue(set(obj_param_dict.keys()).issuperset(param_name_set))
-
-        self.assertEqual(obj_param_dict["--arg3"].short_summary, "Arg 3's json summary.")
-        self.assertEqual(obj_param_dict["--arg3"].long_summary, "A truly true description of this parameter.")
-        self.assertEqual(obj_param_dict["--arg3"].value_sources[0], {"string": "Number range: 0 to 10"})
-        self.assertEqual(obj_param_dict["--arg3"].value_sources[1]['link'],
-                         {"url": "https://www.foo-json.com", "title": "foo-json"})
-        self.assertEqual(obj_param_dict["--arg3"].value_sources[2]['link'],
-                         {"command": "az test show", "title": "Show test details. Json file"})
-
-        self.assertEqual(command_help_obj.examples[0].short_summary, "A simple example from json")
-        self.assertEqual(command_help_obj.examples[0].long_summary, "More detail on the simple example.")
-        self.assertEqual(command_help_obj.examples[0].command, "az test alpha --arg1 alpha --arg2 beta --arg3 chi")
-        self.assertEqual(command_help_obj.examples[0].supported_profiles, "2018-03-01-hybrid, latest")
-
-        # validate other parameters, which have help from help.py and help.yamls
-        self.assertEqual(obj_param_dict["--arg1 -a"].short_summary, "A short summary.")
-        self.assertEqual(obj_param_dict["--arg2 -b"].short_summary, "Arg 2's summary.")
-        self.assertEqual(obj_param_dict["ARG4"].short_summary, "Arg4's summary, yaml. Positional arg, not required.")
-        # arg2's help from help.yaml still preserved.
-        self.assertEqual(obj_param_dict["--arg2 -b"].long_summary, "A true description of this parameter.")
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[0], {"string": "Number range: -5.0 to 5.0"})
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[1]['link'], {"url": "https://www.foo.com",
-                                                                                "title": "foo"})
-        self.assertEqual(obj_param_dict["--arg2 -b"].value_sources[2]['link'], {"command": "az test show",
-                                                                                "title": "Show test details"})
-
-    # create a temporary file in the temp dir. Return the path of the file.
-    def _create_new_temp_file(self, data, suffix=""):
-        with tempfile.NamedTemporaryFile(mode='w', dir=self._tempdirName, delete=False, suffix=suffix) as f:
-            f.write(data)
-            return f.name
 
 
 class TestHelpSupportedProfiles(unittest.TestCase):
